Add files via upload
The pipeline shows all the code necessary to recreate my work. The final data shows the data frame that includes the 10 topic timelines. Ultimate files is used as a reference to the files used in the data cleaning steps. The clean text file is to get an idea of what a cleaned data file looks like. More data can be found on Google Drive, linked in the thesis.